// SDKs/twilio.com/twilio_chat_v2/1.36.0/typescript/src/sdk/models/shared/chatv2servicechannelmessage.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class ChatV2ServiceChannelMessage extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=attributes" })
  attributes?: string;

  @Metadata({ data: "json, name=body" })
  body?: string;

  @Metadata({ data: "json, name=channel_sid" })
  channelSid?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=from" })
  from?: string;

  @Metadata({ data: "json, name=index" })
  index?: number;

  @Metadata({ data: "json, name=last_updated_by" })
  lastUpdatedBy?: string;

  @Metadata({ data: "json, name=media" })
  media?: any;

  @Metadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=to" })
  to?: string;

  @Metadata({ data: "json, name=type" })
  type?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;

  @Metadata({ data: "json, name=was_edited" })
  wasEdited?: boolean;
}